Fix initialization bug in `FakeImageService`
When adding this functionality in
https://github.com/kubernetes/kubernetes/pull/88372, I forgot to
allocate a map for the `Errors` field when constructing the object. As a
result, trying to actually use the `InjectError` method failed (as I
noticed when I started trying to write tests to actually use
`InjectError`). Fortunately, `FakeImageService` is only used in tests...
but still, we should fix this issue.

Fixing in a separate diff from the one which will add additional test
coverage (and actually use `InjectError`, because I don't like having
non-working code in master.)

We could also revert the original commit and then re-merge with this
fix, but that seems like unnecessary work given we already have a fix
ready to go.

Kubernetes-commit: 5362e8b6cb1ac42356beab84a6ee8db3ccb495b5