Add `image_id` to CRI `ContainerStatus` message
There is a conversion function `ConvertPodStatusToRunningPod`, which
can override the `Container.ImageID` into a digested reference from the
`ContainerStatus` CRI RPC, which gets mapped from the `image_ref`:

https://github.com/kubernetes/kubernetes/blob/411c29c39f03687a30a8667295c61590def8fc89/pkg/kubelet/container/helpers.go#L259-L292

To avoid that failure case, we now introduce the same `image_id` into
the container status and let runtimes separate the fields.

We also add a note that the mapping from the digested reference of the
CRI to the Kubernetes Pod API `ImageID` field is intentional and should
not change.

Follow-up on: https://github.com/kubernetes/kubernetes/pull/123508

Signed-off-by: Sascha Grunert <sgrunert@redhat.com>

Kubernetes-commit: e38531e9a2359c2ba1505cb04d62d6810edc616e